Circle hover solution
Added a script to main, inside _input() when the mouse is moved, I use intersect_point() to get an Array[Dictionary] that contains all overlapping areas.

It seems like the order is randomized though, so I run a sort on the resulting array to sort based on the tree order so that the result is always consistent.

Finally there is a setter property that stores the currently hovered circle, which then asks the old circle to shrink and the new circle to grow.
- The setter has some checks to run logic based on what the previous value was and what the new value is